+++ src/Exception/InvalidMethodException.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Kuria\Enum\Exception;
+
+class InvalidMethodException extends \BadMethodCallException implements ExceptionInterface
+{
+}