+++ tests/TestSubject/StringEnumObject.php
@@ -2,12 +2,12 @@
 
 namespace Kuria\Enum\TestSubject;
 
-use Kuria\Enum\Enum;
+use Kuria\Enum\EnumObject;
 
 /**
  * @internal
  */
-abstract class StringEnum extends Enum
+class StringEnumObject extends EnumObject
 {
     const NUMERIC_STRING_KEY = '123';
     const EMPTY_STRING_KEY = '';